rename jobs to xjobs
