+++ xjobs/test/test_jobs_utils.py
@@ -1,4 +1,4 @@
-from jobs import utils
+from xjobs import utils
 import pandas as pd
 from pandas.testing import assert_frame_equal
 from pandas import Timestamp
